requirements.txt
flask==2.3.3
pandas==2.1.0
pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
numpy==1.24.3
scikit-learn==1.3.0
opencv-python==4.8.0.76
werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
markupsafe==2.1.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
cycler==0.11.0
kiwisolver==1.4.4
pyparsing==3.0.9   # ✅ hạ xuống tương thích với matplotlib
fonttools==4.42.1
packaging==23.1
contourpy==1.1.0
gunicorn==21.2.0
scikit-learn==1.3.0
plotly==5.18.0